SYNC-1622 Engine - Move packages
