+++ modules/apps/sync/sync-engine/src/main/java/com/liferay/sync/engine/lan/server/file/LanFileServerInitializer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sync.engine.lan.fileserver;
+package com.liferay.sync.engine.lan.server.file;
 
 import com.liferay.sync.engine.lan.util.LanClientUtil;
 import com.liferay.sync.engine.lan.util.LanPEMParserUtil;
