+++ modules/apps/sync/sync-engine/src/main/java/com/liferay/sync/engine/lan/session/LanSession.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sync.engine.lan;
+package com.liferay.sync.engine.lan.session;
 
 import com.liferay.sync.engine.document.library.handler.Handler;
 import com.liferay.sync.engine.lan.util.LanClientUtil;
