+++ modules/apps/sync/sync-engine/src/main/java/com/liferay/sync/engine/lan/session/NoSuchSyncLanClientException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sync.engine.lan;
+package com.liferay.sync.engine.lan.session;
 
 /**
  * @author Dennis Ju
